Add additional buzz patterns
This buzz also changes how patterns are handled. Instead of having a few
patterns hardcoded, it can now handle any pattern string, which contains
'.' and '_', where the '.' is a short buzz and the '_' is a long
buzz. This opens the way for a feature like letting the user choose
his/her own custom pattern.